requirements.txt
fastapi
uvicorn[standard]
requests
aiofiles
python-dotenv
